refactor: Select 파일 및 타입 분리
